refactor: refactor js modules
