+++ index.html
@@ -8,21 +8,7 @@
   <title>Virtual keyboard</title>
 </head>
 <body class="body">
-  <!-- <div class="wrapper">
-    <h1 class="header__title">Virtual keyboard</h1>
-    <main class="main">
-      <textarea class="main__output" rows="5" value=""></textarea>
-      <section class="main__keyboard keyboard">
 
-          <button class="keyboard__key">a</button>
-          <button class="keyboard__key">b</button>
-          <br>
-          <button class="keyboard__key keyboard__key_big">delete</button>
-          <button class="keyboard__key keyboard__key_extra-big">space</button>
-
-      </section>
-    </main>
-  </div> -->
   <script type="module" src="./script/app.js"></script>
 </body>
 </html>
