+++ script/consts.js
@@ -1,0 +1,27 @@
+export const CssClasses = {
+  BODY: 'body',
+  WRAPPER: 'wrapper',
+  MAIN: 'main',
+  TEXTAREA: 'main__output',
+  HEADER: 'header__title',
+  DISCLAIMER: 'main__disclaimer',
+  KEYBOARD: 'keyboard',
+  KEY: 'keyboard__key',
+  KEY_MEDIUM: 'keyboard__key_med',
+  KEY_BIG: 'keyboard__key_big',
+  KEY_EXTRA_BIG: 'keyboard__key_extra-big',
+  KEY_DARK: 'keyboard__key_dark',
+  KEY_BLANK: 'keyboard__key_blank',
+  KEY_CAPS: 'keyboard__key_caps',
+  KEYBOARD_ARROW: 'keyboard__arrow',
+  KEY_ARROW: 'key__arrow',
+  KEY_ARROW_UP: 'key__arrow_up',
+  KEY_ARROW_RIGHT: 'key__arrow_right',
+  KEY_ARROW_LEFT: 'key__arrow_left',
+  KEY_ARROW_DOWN: 'key__arrow_down',
+};
+
+export const KeyboardConst = {
+  CHANGE_LANG_BUTTON1: 'ControlLeft',
+  CHANGE_LANG_BUTTON2: 'AltLeft',
+};
