+++ script/localStorage.js
@@ -1,9 +1,7 @@
-export function setLocalStorage(Obj) {
-  localStorage.setItem('lang', Obj.output.lang);
+export function setLocalStorage(name, item) {
+  localStorage.setItem(name, item);
 }
 
-export function getLocalStorage(Obj) {
-  if (localStorage.getItem('lang')) {
-    Obj.output.lang = localStorage.getItem('lang');
-  }
+export function getLocalStorage(name) {
+      return localStorage.getItem(name);
 }
