+++ style.css
@@ -30,7 +30,6 @@ button {
 
 .main__output {
   width: 100%;
-  /* height: 15vh; */
   border: 1px solid rgba(216, 103, 37, 0.5);
   border-radius: 5px;
   padding: 10px;
@@ -56,7 +55,7 @@ button {
   border-radius: 5px;
   padding: 5px;
   color: white;
-  transition: all ease-in .3s;
+  transition: all .3s;
 }
 
 .keyboard__key:hover {
